+++ packages/zeebe-element-templates-json-schema/test/fixtures/form-definition-missing-zeebe-user-task.js
@@ -1,0 +1,81 @@
+export const template = {
+  '$schema': 'https://unpkg.com/@camunda/zeebe-element-templates-json-schema/resources/schema.json',
+  'name': 'Form Definition with external reference',
+  'id': 'formDefinitionWithExternalReference',
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'elementType': {
+    'value': 'bpmn:UserTask'
+  },
+  'properties': [
+    {
+      'type': 'Hidden',
+      'value': 'aReference',
+      'binding': {
+        'type': 'zeebe:formDefinition',
+        'property': 'externalReference'
+      }
+    }
+  ]
+};
+
+export const errors = [
+  {
+    keyword: 'errorMessage',
+    dataPath: '',
+    schemaPath: '#/allOf/1/allOf/4/then/allOf/0/errorMessage',
+    params: {
+      errors: [
+        {
+          keyword: 'const',
+          dataPath: '/properties/0/binding/type',
+          schemaPath: '#/allOf/1/allOf/4/then/allOf/0/properties/properties/contains/properties/binding/properties/type/const',
+          params: {
+            allowedValue: 'zeebe:userTask'
+          },
+          message: 'should be equal to constant',
+          emUsed: true
+        },
+        {
+          keyword: 'contains',
+          dataPath: '/properties',
+          schemaPath: '#/allOf/1/allOf/4/then/allOf/0/properties/properties/contains',
+          params: {
+            minContains: 1
+          },
+          message: 'should contain at least 1 valid item(s)',
+          emUsed: true
+        }
+      ]
+    },
+    message: 'When using "zeebe:formDefinition", "zeebe:userTask" must be set on the same element'
+  },
+  {
+    keyword: 'if',
+    dataPath: '',
+    schemaPath: '#/allOf/1/allOf/4/if',
+    params: {
+      failingKeyword: 'then'
+    },
+    message: 'should match "then" schema'
+  },
+  {
+    keyword: 'type',
+    dataPath: '',
+    schemaPath: '#/oneOf/1/type',
+    params: {
+      type: 'array'
+    },
+    message: 'should be array'
+  },
+  {
+    keyword: 'oneOf',
+    dataPath: '',
+    schemaPath: '#/oneOf',
+    params: {
+      passingSchemas: null
+    },
+    message: 'should match exactly one schema in oneOf'
+  }
+];