+++ packages/zeebe-element-templates-json-schema/test/fixtures/form-definition-with-formId.js
@@ -1,0 +1,29 @@
+export const template = {
+  '$schema': 'https://unpkg.com/@camunda/zeebe-element-templates-json-schema/resources/schema.json',
+  'name': 'Form Definition with FormId',
+  'id': 'formDefinitionWithFormId',
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'elementType': {
+    'value': 'bpmn:UserTask'
+  },
+  'properties': [
+    {
+      'type': 'Hidden',
+      'binding': {
+        'type': 'zeebe:userTask',
+      }
+    },
+    {
+      'type': 'Hidden',
+      'value': 'aFormId',
+      'binding': {
+        'type': 'zeebe:formDefinition',
+        'property': 'formId'
+      }
+    }
+  ]
+};
+
+export const errors = null;